feat: implemented atomic design structure, ui for hero header
