--- src/components/Footer/index.js
@@ -1,14 +1,0 @@
-import React, { Component } from "react";
-import SocialIcons from "../SocialIcons";
-
-class Footer extends Component {
-  render() {
-    return (
-      <div className="footer">
-        <SocialIcons iconWidth="40px" iconColor="#629AFE" />
-      </div>
-    );
-  }
-}
-
-export default Footer;
