+++ src/components/atoms/Button/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import StyledButton from "./styled";
+
+const Button = ({ children, type }) => {
+  return <StyledButton className={type}>{children}</StyledButton>;
+};
+
+export default Button;
