+++ src/components/atoms/SectionWrapper/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import { StyledSectionWrapper } from "./styled";
+
+const SectionWrapper = ({ children }) => {
+  return <StyledSectionWrapper>{children}</StyledSectionWrapper>;
+};
+
+export default SectionWrapper;
