+++ src/components/atoms/SectionWrapper/styled.js
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+export const StyledSectionWrapper = styled.section`
+  max-width: 1200px;
+  margin: auto;
+  width: 100%;
+`;
