+++ src/components/molecules/SocialIcons/styled.js
@@ -1,0 +1,15 @@
+import styled from "styled-components";
+
+export const StyledSocialIcons = styled.div`
+  display: flex;
+  align-items: center;
+
+  svg {
+    width: ${props => props.iconWidth};
+    height: auto;
+    margin: 8px;
+    path {
+      fill: ${props => props.iconColor};
+    }
+  }
+`;
