+++ src/components/organisms/Footer/index.js
@@ -1,0 +1,14 @@
+import React, { Component } from "react";
+import SocialIcons from "../../molecules/SocialIcons";
+
+class Footer extends Component {
+  render() {
+    return (
+      <div className="footer">
+        <SocialIcons iconWidth="40px" iconColor="#629AFE" />
+      </div>
+    );
+  }
+}
+
+export default Footer;
