+++ src/components/pages/Home/index.js
@@ -1,0 +1,8 @@
+import React from "react";
+import HomeHero from "../../templates/Home/Hero";
+
+const Home = () => {
+  return <HomeHero />;
+};
+
+export default Home;
