+++ src/components/pages/JobBoard.js
@@ -1,0 +1,15 @@
+import React, { Component } from 'react'
+
+class JobBoard extends Component {
+
+    render() {
+
+        return (
+            <div className="main-content">
+              <p>navigated to: JOB-BOARD</p>
+            </div>
+        )
+      }
+    }
+
+export default JobBoard
