--- src/pages/Home.js
@@ -1,11 +1,0 @@
-import React, { Component } from "react";
-
-const Home: Component = () => {
-  return (
-    <div className="main-content">
-      <p>Home</p>
-    </div>
-  );
-};
-
-export default Home;
