--- src/pages/JobBoard.js
@@ -1,15 +1,0 @@
-import React, { Component } from 'react'
-
-class JobBoard extends Component {
-
-    render() {
-
-        return (
-            <div className="main-content">
-              <p>navigated to: JOB-BOARD</p>
-            </div>
-        )
-      }
-    }
-
-export default JobBoard
